rke2-uninstall.sh enhancements
Fix a couple gaps/oversights in uninstall. First don't remove /etc/rancher/rke2 unconditionally so that if a user created config.yaml, or any other file is present we don't remove user data that may not be able to be recreated. Aka a hand crafted config.yaml file that isn't backed up shouldn't be removed.

Secondarily, check if /var/lib/rancher is a mountpoint or not, if it is do not try to remove it with rm.

+++ bundle/bin/rke2-uninstall.sh
@@ -44,12 +44,15 @@ uninstall_remove_files()
     rm -f "${INSTALL_RKE2_ROOT}/bin/rke2"
     rm -f "${INSTALL_RKE2_ROOT}/bin/rke2-killall.sh"
     rm -rf "${INSTALL_RKE2_ROOT}/share/rke2"
-    rm -rf /etc/rancher/rke2
+    rm -f /etc/rancher/rke2/rke2.yaml
+    rm -d /etc/rancher/rke2 || true
     rm -rf /etc/rancher/node
     rm -d /etc/rancher || true
     rm -rf /var/lib/kubelet
     rm -rf /var/lib/rancher/rke2
-    rm -d /var/lib/rancher || true
+    if ! mountpoint /var/lib/rancher > /dev/null 2>&1; then
+        rm -d /var/lib/rancher || true
+    fi
 }
 
 uninstall_remove_self()
